style: add banner thumbs for banner item
