+++ app/routes/anime/discover.tsx
@@ -3,22 +3,22 @@ import { json, LoaderFunction, DataFunctionArgs, MetaFunction } from '@remix-run
 import { Container } from '@nextui-org/react';
 import { motion } from 'framer-motion';
 
-import { getAnimeTrending } from '~/services/consumet/anilist/anilist.server';
-import { authenticate } from '~/services/supabase';
 import AnimeList from '~/src/components/anime/AnimeList';
+import { authenticate } from '~/services/supabase';
+import { getAnimeAdvancedSearch } from '~/services/consumet/anilist/anilist.server';
 
 type LoaderData = {
-  items: Awaited<ReturnType<typeof getAnimeTrending>>;
+  items: Awaited<ReturnType<typeof getAnimeAdvancedSearch>>;
 };
 
 export const meta: MetaFunction = () => ({
-  title: 'Watch Top Trending anime free | Sora',
+  title: 'Discover and watch anime for free | Sora',
   description:
     'Official Sora website to watch anime online HD for free, Watch TV show & TV series and Download all anime FREE',
   keywords:
     'watch free anime, free anime to watch online, watch anime online free, free anime streaming, free anime full, free anime download, watch anime hd, anime to watch',
-  'og:url': 'https://sora-anime.vercel.app/anime/trending',
-  'og:title': 'Watch Top Trending anime free | Sora',
+  'og:url': 'https://sora-anime.vercel.app/anime/discover',
+  'og:title': 'Discover and watch anime for free | Sora',
   'og:description':
     'Official Sora website to watch anime online HD for free, Watch TV show & TV series and Download all anime FREE',
 });
@@ -29,21 +29,57 @@ export const loader: LoaderFunction = async ({ request }: DataFunctionArgs) => {
   const url = new URL(request.url);
   let page = Number(url.searchParams.get('page'));
   if (!page && (page < 1 || page > 1000)) page = 1;
+  const query = url.searchParams.get('query') || undefined;
+  const type = (url.searchParams.get('type') as 'ANIME' | 'MANGA') || 'ANIME';
+  const season =
+    (url.searchParams.get('season') as 'WINTER' | 'SPRING' | 'SUMMER' | 'FALL') || undefined;
+  const format =
+    (url.searchParams.get('format') as
+      | 'TV'
+      | 'TV_SHORT'
+      | 'OVA'
+      | 'ONA'
+      | 'MOVIE'
+      | 'SPECIAL'
+      | 'MUSIC') || undefined;
+  const sort = url.searchParams.get('sort')?.split(',') || undefined;
+  const status =
+    (url.searchParams.get('status') as
+      | 'RELEASING'
+      | 'NOT_YET_RELEASED'
+      | 'FINISHED'
+      | 'CANCELLED'
+      | 'HIATUS') || undefined;
+  const genres = url.searchParams.get('genres')?.split(',') || undefined;
+  const id = url.searchParams.get('id') || undefined;
+  const year = Number(url.searchParams.get('number')) || undefined;
 
   return json<LoaderData>({
-    items: await getAnimeTrending(page, 20),
+    items: await getAnimeAdvancedSearch(
+      query,
+      type,
+      page,
+      20,
+      season,
+      format,
+      sort,
+      genres,
+      id,
+      year,
+      status,
+    ),
   });
 };
 
 export const handle = {
   breadcrumb: () => (
-    <Link to="/anime/trending" aria-label="Trending Anime">
-      Trending Anime
+    <Link to="/anime/discover" aria-label="Discover Anime">
+      Discover Anime
     </Link>
   ),
 };
 
-const TrendingAnime = () => {
+const DiscoverAnime = () => {
   const { items } = useLoaderData<LoaderData>();
   const location = useLocation();
 
@@ -73,8 +109,8 @@ const TrendingAnime = () => {
             listType="grid"
             items={items.results}
             hasNextPage={items.hasNextPage || false}
-            listName="Trending Anime"
-            routeName="/anime/trending"
+            listName="Discover Anime"
+            routeName={location.pathname}
             virtual
           />
         )}
@@ -83,4 +119,4 @@ const TrendingAnime = () => {
   );
 };
 
-export default TrendingAnime;
+export default DiscoverAnime;
