+++ app/routes/collections/$collectionsId.tsx
@@ -85,8 +85,8 @@ const CollectionDetail = () => {
         alignItems="center"
         css={{
           '@xsMax': {
-            paddingLeft: 'calc(var(--nextui-space-sm))',
-            paddingRight: 'calc(var(--nextui-space-sm))',
+            paddingLeft: '$sm',
+            paddingRight: '$sm',
             paddingBottom: '65px',
           },
         }}
