+++ app/routes/index.tsx
@@ -103,8 +103,8 @@ const Index = () => {
           minHeight: '564px',
           padding: 0,
           '@xsMax': {
-            paddingLeft: 'calc(var(--nextui-space-sm))',
-            paddingRight: 'calc(var(--nextui-space-sm))',
+            paddingLeft: '$sm',
+            paddingRight: '$sm',
           },
         }}
       >
