+++ app/routes/movies/$movieId/photos.tsx
@@ -102,8 +102,8 @@ const PhotosPage = () => {
       css={{
         flexDirection: 'column',
         '@xsMax': {
-          paddingLeft: 'calc(var(--nextui-space-sm))',
-          paddingRight: 'calc(var(--nextui-space-sm))',
+          paddingLeft: '$sm',
+          paddingRight: '$sm',
         },
         '@xs': {
           paddingLeft: '88px',
