+++ app/routes/movies/$movieId/recommendations.tsx
@@ -71,8 +71,8 @@ const RecommendationsPage = () => {
       css={{
         flexDirection: 'column',
         '@xsMax': {
-          paddingLeft: 'calc(var(--nextui-space-sm))',
-          paddingRight: 'calc(var(--nextui-space-sm))',
+          paddingLeft: '$sm',
+          paddingRight: '$sm',
         },
         '@xs': {
           paddingLeft: '88px',
