+++ app/routes/movies/discover.tsx
@@ -164,8 +164,8 @@ const ListMovies = () => {
         alignItems="center"
         css={{
           '@xsMax': {
-            paddingLeft: 'calc(var(--nextui-space-sm))',
-            paddingRight: 'calc(var(--nextui-space-sm))',
+            paddingLeft: '$sm',
+            paddingRight: '$sm',
           },
         }}
       >
