+++ app/routes/movies/index.tsx
@@ -150,8 +150,8 @@ const MoviesIndexPage = () => {
           marginTop: '48px',
           minHeight: '564px',
           '@xsMax': {
-            paddingLeft: 'calc(var(--nextui-space-sm))',
-            paddingRight: 'calc(var(--nextui-space-sm))',
+            paddingLeft: '$sm',
+            paddingRight: '$sm',
           },
         }}
       >
