+++ app/routes/search/anime/$animeKeyword.tsx
@@ -71,8 +71,8 @@ const SearchRoute = () => {
         alignItems="center"
         css={{
           '@xsMax': {
-            paddingLeft: 'calc(var(--nextui-space-sm))',
-            paddingRight: 'calc(var(--nextui-space-sm))',
+            paddingLeft: '$sm',
+            paddingRight: '$sm',
           },
         }}
       >
