+++ app/routes/search/anime/index.tsx
@@ -56,8 +56,8 @@ const SearchRoute = () => {
         alignItems="center"
         css={{
           '@xsMax': {
-            paddingLeft: 'calc(var(--nextui-space-sm))',
-            paddingRight: 'calc(var(--nextui-space-sm))',
+            paddingLeft: '$sm',
+            paddingRight: '$sm',
           },
         }}
       >
