+++ app/routes/search/people/$peopleKeyword.tsx
@@ -83,8 +83,8 @@ const SearchRoute = () => {
         alignItems="center"
         css={{
           '@xsMax': {
-            paddingLeft: 'calc(var(--nextui-space-sm))',
-            paddingRight: 'calc(var(--nextui-space-sm))',
+            paddingLeft: '$sm',
+            paddingRight: '$sm',
           },
         }}
       >
