+++ app/routes/tv-shows/$tvId/crew.tsx
@@ -41,8 +41,8 @@ const CrewPage = () => {
       css={{
         flexDirection: 'column',
         '@xsMax': {
-          paddingLeft: 'calc(var(--nextui-space-sm))',
-          paddingRight: 'calc(var(--nextui-space-sm))',
+          paddingLeft: '$sm',
+          paddingRight: '$sm',
         },
         '@xs': {
           paddingLeft: '88px',
