+++ app/routes/tv-shows/on-tv.tsx
@@ -80,8 +80,8 @@ const ListTvShows = () => {
         justify="center"
         css={{
           '@xsMax': {
-            paddingLeft: 'calc(var(--nextui-space-sm))',
-            paddingRight: 'calc(var(--nextui-space-sm))',
+            paddingLeft: '$sm',
+            paddingRight: '$sm',
           },
         }}
       >
