+++ app/src/components/layouts/Header.tsx
@@ -1,21 +1,6 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import * as React from 'react';
-import { Link, useLocation, useNavigate } from '@remix-run/react';
-import {
-  Avatar,
-  Button,
-  Grid,
-  Row,
-  Switch,
-  Tooltip,
-  Popover,
-  useTheme,
-  styled,
-  Divider,
-  Spacer,
-} from '@nextui-org/react';
-import { useTheme as useRemixTheme } from 'next-themes';
-import { motion } from 'framer-motion';
+import { Button, Grid, Row, Tooltip, Popover, Spacer } from '@nextui-org/react';
 import type { User } from '@supabase/supabase-js';
 import { Player } from '@lottiefiles/react-lottie-player';
 import type { AnimationItem } from 'lottie-web';
@@ -24,22 +9,20 @@ import { useTranslation } from 'react-i18next';
 import useMediaQuery from '~/hooks/useMediaQuery';
 import useScrollDirection from '~/hooks/useScrollDirection';
 
+import { pages, searchDropdown } from '~/src/constants/navPages';
+
 /* Components */
-import { H5, H6 } from '~/src/components/styles/Text.styles';
 import NavLink from '../elements/NavLink';
+import MultiLevelDropdown from './MultiLevelDropdown';
+import { AppBar } from './Layout.styles';
 
 /* Assets */
-import kleeCute from '../../assets/images/avatar.png';
-import SunIcon from '../../assets/icons/SunIcon.js';
-import MoonIcon from '../../assets/icons/MoonIcon.js';
 import MenuIcon from '../../assets/icons/MenuIcon.js';
 import SearchIcon from '../../assets/icons/SearchIcon.js';
-import GlobalIcon from '../../assets/icons/GlobalIcon.js';
 // import menuNavBlack from '../../assets/lotties/lottieflow-menu-nav-11-6-000000-easey.json';
 // import menuNavWhite from '../../assets/lotties/lottieflow-menu-nav-11-6-FFFFFF-easey.json';
 // import arrowLeftBlack from '../../assets/lotties/lottieflow-arrow-08-1-000000-easey.json';
 // import arrowLeftWhite from '../../assets/lotties/lottieflow-arrow-08-1-FFFFFF-easey.json';
-import arrowLeft from '../../assets/lotties/lottieflow-arrow-08-1-0072F5-easey.json';
 import dropdown from '../../assets/lotties/lottieflow-dropdown-03-0072F5-easey.json';
 
 interface IHeaderProps {
@@ -52,71 +35,6 @@ interface IHeaderProps {
 export const handle = {
   i18n: 'header',
 };
-
-const pages = [
-  {
-    pageName: 'movies',
-    pageLink: 'movies',
-    pageDropdown: [
-      { pageName: 'Discover', pageLink: 'movies/discover' },
-      { pageName: 'popular', pageLink: 'movies/popular' },
-      { pageName: 'topRated', pageLink: 'movies/top-rated' },
-      { pageName: 'upcoming', pageLink: 'movies/upcoming' },
-    ],
-  },
-  {
-    pageName: 'tv',
-    pageLink: 'tv-shows',
-    pageDropdown: [
-      { pageName: 'Discover', pageLink: 'tv-shows/discover' },
-      { pageName: 'popular', pageLink: 'tv-shows/popular' },
-      { pageName: 'topRated', pageLink: 'tv-shows/top-rated' },
-      { pageName: 'onTv', pageLink: 'tv-shows/on-tv' },
-    ],
-  },
-  {
-    pageName: 'Anime',
-    pageLink: 'anime',
-    pageDropdown: [
-      // { pageName: 'Discover', pageLink: 'anime/discover' },
-      { pageName: 'Recent Episodes', pageLink: 'anime/recent-episodes' },
-      { pageName: 'Popular Anime', pageLink: 'anime/popular' },
-      { pageName: 'Trending Anime', pageLink: 'anime/trending' },
-    ],
-  },
-];
-
-const searchDropdown = [
-  { pageName: 'search.title.movie', pageLink: 'search/movie' },
-  { pageName: 'search.title.tv', pageLink: 'search/tv' },
-  { pageName: 'search.title.people', pageLink: 'search/people' },
-  { pageName: 'search.title.anime', pageLink: 'search/anime' },
-];
-
-const slideHorizontalAnimation = {
-  left: {
-    x: 0,
-    transition: {
-      duration: 0.3,
-    },
-  },
-  right: {
-    x: -290,
-    transition: {
-      duration: 0.3,
-    },
-  },
-};
-
-const languages = ['en', 'fr', 'vi'];
-
-const AppBar = styled(Grid.Container, {
-  zIndex: 990,
-  position: 'sticky',
-  '@sm': {
-    borderTopLeftRadius: '$xl',
-  },
-});
 
 const DropdownPage = ({
   pagesDropdown,
@@ -144,346 +62,6 @@ const DropdownPage = ({
         </>
       ))}
     </Grid.Container>
-  );
-};
-
-const MultiLevelDropdown = ({ user }: { user: User | undefined }) => {
-  const { isDark } = useTheme();
-  const { setTheme } = useRemixTheme();
-  const navigate = useNavigate();
-  const [isLeftMenu, setIsLeftMenu] = React.useState(true);
-  const [isLanguageTab, setIsLanguageTab] = React.useState(false);
-  const [isDisplayTab, setIsDisplayTab] = React.useState(false);
-  const { t } = useTranslation('header');
-
-  const location = useLocation();
-
-  return (
-    <motion.div
-      className="dropdown"
-      initial="left"
-      animate={isLeftMenu ? 'left' : 'right'}
-      variants={slideHorizontalAnimation}
-      style={{
-        display: 'flex',
-        flexDirection: 'row',
-        alignItems: 'flex-start',
-        height: '100%',
-        position: 'relative',
-        transition: 'height 0.3s',
-        width: '590px',
-        transform: `${isLeftMenu ? 'none' : ' translateZ(0px) translateX(-290px)'}`,
-      }}
-    >
-      <motion.div>
-        <Grid.Container
-          css={{
-            flexDirection: 'column',
-          }}
-        >
-          <Grid css={{ margin: '10px 0 0 10px', width: 280, minHeight: 65, display: 'block' }}>
-            <Button
-              light
-              color="primary"
-              size="md"
-              css={{ w: 260, h: 50 }}
-              icon={
-                <Avatar
-                  size="md"
-                  alt="Klee Cute"
-                  src={kleeCute}
-                  color="primary"
-                  bordered
-                  css={{ cursor: 'pointer' }}
-                />
-              }
-            >
-              {user ? (
-                <H6 h6 weight="bold" color="inherit" css={{ marginLeft: '3.5rem !important' }}>
-                  {user?.email ?? 'klee@example.com'}
-                </H6>
-              ) : (
-                <Link to={`/sign-in?ref=${location.pathname ?? ''}`}>
-                  <H6 h6 weight="bold" color="inherit" css={{ textTransform: 'uppercase' }}>
-                    Sign In
-                  </H6>
-                </Link>
-              )}
-            </Button>
-            <Divider x={1} css={{ width: 260, margin: '10px 40px 0 0' }} />
-          </Grid>
-          <Grid css={{ margin: '10px 0 0 10px', width: 280, minHeight: 65, display: 'block' }}>
-            <Button
-              flat
-              color="primary"
-              size="md"
-              onClick={() => {
-                setIsLanguageTab(true);
-                setIsLeftMenu(false);
-              }}
-              css={{ w: 260, h: 50 }}
-              icon={<GlobalIcon />}
-            >
-              <H6 h6 color="primary">
-                Language
-              </H6>
-            </Button>
-          </Grid>
-          <Grid css={{ margin: '10px 0 0 10px', width: 280, minHeight: 65, display: 'block' }}>
-            <Button
-              flat
-              color="primary"
-              size="md"
-              onClick={() => {
-                setIsDisplayTab(true);
-                setIsLeftMenu(false);
-              }}
-              css={{ w: 260, h: 50 }}
-            >
-              <H6 h6 color="primary">
-                Display
-              </H6>
-            </Button>
-          </Grid>
-          <Grid css={{ margin: '10px 0 0 10px', width: 280, minHeight: 65, display: 'block' }}>
-            {user ? (
-              <Button flat color="error" size="md" css={{ w: 260, h: 50 }}>
-                <Link to={`/sign-out?ref=${location.pathname ?? ''}`}>
-                  <H5 h5 color="error">
-                    Log out
-                  </H5>
-                </Link>
-              </Button>
-            ) : (
-              <Button flat color="primary" size="md" css={{ w: 260, h: 50 }}>
-                <Link to={`/sign-up?ref=${location.pathname ?? ''}`}>Sign Up</Link>
-              </Button>
-            )}
-          </Grid>
-        </Grid.Container>
-      </motion.div>
-      <motion.div>
-        <Grid.Container
-          css={{
-            flexDirection: 'column',
-          }}
-        >
-          {isLanguageTab && (
-            <>
-              <Grid css={{ margin: '10px 0 0 10px', width: 280, minHeight: 65, display: 'block' }}>
-                <Button
-                  light
-                  color="primary"
-                  size="md"
-                  onClick={() => {
-                    setIsLanguageTab(false);
-                    setIsLeftMenu(true);
-                  }}
-                  css={{ w: 260, h: 50 }}
-                  icon={
-                    <Player
-                      src={arrowLeft}
-                      hover
-                      autoplay={false}
-                      speed={0.75}
-                      className="w-8 h-8"
-                      loop
-                    />
-                  }
-                >
-                  <H6 h6 color="primary">
-                    Language
-                  </H6>
-                </Button>
-                <Divider x={1} css={{ width: 260, margin: '10px 40px 0 0' }} />
-              </Grid>
-              {languages.map((lng) => (
-                <Grid
-                  key={lng}
-                  css={{ margin: '10px 0 0 10px', width: 280, minHeight: 65, display: 'block' }}
-                >
-                  <Button
-                    flat
-                    color="primary"
-                    size="md"
-                    onClick={() => {
-                      setIsLanguageTab(false);
-                      setIsLeftMenu(true);
-                      navigate(`${location.pathname}?lng=${lng}`);
-                    }}
-                    css={{ w: 260, h: 50 }}
-                  >
-                    <H6 h6 color="primary">
-                      {t(lng)}
-                    </H6>
-                  </Button>
-                </Grid>
-              ))}
-            </>
-          )}
-          {isDisplayTab && (
-            <>
-              <Grid css={{ margin: '10px 0 0 10px', width: 280, minHeight: 65, display: 'block' }}>
-                <Button
-                  light
-                  color="primary"
-                  size="md"
-                  onClick={() => {
-                    setIsDisplayTab(false);
-                    setIsLeftMenu(true);
-                  }}
-                  css={{ w: 260, h: 50 }}
-                  icon={
-                    <Player
-                      src={arrowLeft}
-                      hover
-                      autoplay={false}
-                      speed={0.75}
-                      className="w-8 h-8"
-                      loop
-                    />
-                  }
-                >
-                  <H6 h6 color="primary">
-                    Display
-                  </H6>
-                </Button>
-                <Divider x={1} css={{ width: 260, margin: '10px 40px 0 0' }} />
-              </Grid>
-              <Grid
-                direction="row"
-                justify="space-around"
-                alignItems="center"
-                css={{
-                  display: 'flex',
-                  width: 280,
-                  minHeight: 65,
-                }}
-              >
-                <H6 h6>Light mode</H6>
-                <Switch
-                  checked={isDark}
-                  size="md"
-                  onChange={(e) => setTheme(e.target.checked ? 'dark' : 'light')}
-                  iconOn={<MoonIcon filled />}
-                  iconOff={<SunIcon filled />}
-                  css={{ padding: 0 }}
-                />
-                <H6 h6>Dark mode</H6>
-              </Grid>
-              <Grid css={{ margin: '10px 0 0 10px', width: 280, minHeight: 65, display: 'block' }}>
-                <Button
-                  flat
-                  size="md"
-                  onClick={() => {
-                    setTheme('light');
-                    setTheme('bumblebee');
-                  }}
-                  css={{
-                    w: 260,
-                    h: 50,
-                    color: '#C08921 !important',
-                    backgroundColor: '#FBEAAB !important',
-                  }}
-                >
-                  Bumblebee Theme
-                </Button>
-              </Grid>
-              <Grid css={{ margin: '10px 0 0 10px', width: 280, minHeight: 65, display: 'block' }}>
-                <Button
-                  flat
-                  size="md"
-                  onClick={() => {
-                    setTheme('dark');
-                    setTheme('synthwave');
-                  }}
-                  css={{
-                    w: 260,
-                    h: 50,
-                    color: '#D427A5 !important',
-                    backgroundColor: '#FEAEC9 !important',
-                  }}
-                >
-                  Synthwave Theme
-                </Button>
-              </Grid>
-              <Grid css={{ margin: '10px 0 0 10px', width: 280, minHeight: 65, display: 'block' }}>
-                <Button
-                  flat
-                  size="md"
-                  onClick={() => {
-                    setTheme('light');
-                    setTheme('retro');
-                  }}
-                  css={{
-                    w: 260,
-                    h: 50,
-                    color: '#CD6C70 !important',
-                    backgroundColor: '#FDE2D7 !important',
-                  }}
-                >
-                  Retro Theme
-                </Button>
-              </Grid>
-              <Grid css={{ margin: '10px 0 0 10px', width: 280, minHeight: 65, display: 'block' }}>
-                <Button
-                  flat
-                  size="md"
-                  onClick={() => {
-                    setTheme('dark');
-                    setTheme('dracula');
-                  }}
-                  css={{
-                    w: 260,
-                    h: 50,
-                    color: '#DB58B0 !important',
-                    backgroundColor: '#FFC9D8 !important',
-                  }}
-                >
-                  Dracula Theme
-                </Button>
-              </Grid>
-              <Grid css={{ margin: '10px 0 0 10px', width: 280, minHeight: 65, display: 'block' }}>
-                <Button
-                  flat
-                  size="md"
-                  onClick={() => {
-                    setTheme('light');
-                    setTheme('autumn');
-                  }}
-                  css={{
-                    w: 260,
-                    h: 50,
-                    color: '#78022C !important',
-                    backgroundColor: '#F39694 !important',
-                  }}
-                >
-                  Autumn Theme
-                </Button>
-              </Grid>
-              <Grid css={{ margin: '10px 0 0 10px', width: 280, minHeight: 65, display: 'block' }}>
-                <Button
-                  flat
-                  size="md"
-                  onClick={() => {
-                    setTheme('dark');
-                    setTheme('night');
-                  }}
-                  css={{
-                    w: 260,
-                    h: 50,
-                    color: '#2894D5 !important',
-                    backgroundColor: '#AFF5FE !important',
-                  }}
-                >
-                  Night Theme
-                </Button>
-              </Grid>
-            </>
-          )}
-        </Grid.Container>
-      </motion.div>
-    </motion.div>
   );
 };
